angular/src/app/services/posts.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpErrorResponse } from '@angular/common/http';
import { Observable, throwError } from 'rxjs';
import { catchError } from 'rxjs/operators';

@Injectable({
  providedIn: 'root'
})
export class PostsService {
  apiURL = 'http://localhost:8080'


  constructor(public http: HttpClient) { }

  // getPosts(id: any): Observable<any> {
  //   console.log(this.apiURL + '/posts/' + id)
  //   return this.http.get(this.apiURL + '/posts/' + id);
  // }

  getPosts(id: any): Observable<any> {
    return this.http.get(this.apiURL + '/posts/' + id).pipe(
      catchError(error => {
        let errorMsg: string;
        if (error.error instanceof ErrorEvent) {
          errorMsg = `Error: ${error.error.message}`;
        } else {
          errorMsg = this.getServerErrorMessage(error);
        }
        return throwError(() => new Error(errorMsg))
      })
    );
  }

  private getServerErrorMessage(error: HttpErrorResponse): string {
    switch (error.status) {
      case 404: {
        return `Not Found: ${error.message}`;
      }
      case 403: {
        return `Access Denied: ${error.message}`;
      }
      case 500: {
        return `Internal Server Error: ${error.message}`;
      }
      default: {
        return `Unknown Server Error: ${error.message}`;
      }
    }
  }
}
